Moved s_initISET into isettools/scripts
I am not quite sure where this belongs.  I would probably not call it a script, as we use it like a function.  But a script it is, just not a script like many of the others.  So I put it into isettools/scripts.  It is the only thing in that directory.  I guess conceptually, that directory should contain scripts that we are treating as functions.  If all the other scripts in the main scripts directory go away, then this would be clear enough.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>